Make "inmem integration test schema" shared with "Postgres integration test schema"
