+++ integration_test/config/complete.postgres.valid.memeo.js
@@ -23,48 +23,7 @@ const datasources = [
   }
 ]
 
-const notesSchema = {
-  id: 2,
-  name: 'default',
-  // language=GraphQL
-  schema: `
-
-      type Profile {
-          id: ID! @isUnique
-          email: String! @isUnique
-          displayName: String!
-          biography: String!
-          avatarUrl: String!
-          memes: [Meme]!
-      }
-
-      type Meme {
-          id: ID! @isUnique
-          photoUrl: String!
-          ownerId: String!
-      }
-
-      type Query {
-          allProfiles:[Profile!]!
-          profile(email: String!):Profile
-          allMemes:[Meme!]!
-      }
-
-      type Mutation {
-          createProfile(email: String!, displayName: String!, biography: String!, avatarUrl: String!):Profile!
-          updateProfile(id: ID!, email: String!, displayName: String!, biography: String!, avatarUrl: String!):Profile
-          deleteProfile(id: ID!):Boolean!
-          createMeme(ownerId: String!, photoUrl: String!):Meme!
-      }
-
-      type Subscription {
-          _: Boolean
-      }
-
-  `,
-  createdAt: time,
-  updatedAt: time
-}
+const schema = require('./complete.valid.memeo.schema.only')
 
 const resolvers = [
   {
@@ -169,10 +128,10 @@ const resolvers = [
 module.exports = {
   up: async (queryInterface, Sequelize) => {
     await queryInterface.bulkInsert('DataSources', datasources, {})
-    await queryInterface.bulkInsert('GraphQLSchemas', [notesSchema], {})
+    await queryInterface.bulkInsert('GraphQLSchemas', [schema], {})
     return queryInterface.bulkInsert('Resolvers', resolvers, {})
   }
 }
 
 // IMPORTANT: please describe the config here. things would be complicated for test maintainers otherwise
-module.exports.description = 'A complex valid config that uses a in-mem data source'
+module.exports.description = 'A complex valid config that uses a Postgres data source'
